show proc COMMAND with -p --threads option
When initially implemented, the --threads option showed
[thread_of_pid_PID] in the COMMAND column instead of the command. This
is because the original data from "ps auxwwwm" only has a dash "-" in
the COMMAND column for threads. However, the [thread_of_pid_PID] isn't
very useful output because we are already showhing the PID in the
second column.

This commit finishes the job. We now show the proper COMMAND for each
thread.